Add Nginx reconciliation to workload controller
This update introduces the reconciliation of an optional NGINX deployment on top of the in-operator simulations. It checks if NGINX deployment is enabled in the workload spec and then ensures the ConfigMap, Secret, HTML ConfigMap and deployment are in the desired state. It also cleans up existing NGINX resources if the deployment is not enabled. The workload's status fields have been extended to hold relevant resource hashes and deployment status information.